Split demo:001 task into two: create course and users and everything else
So that Ansible can run sparfa in between demo:001 and demo:002.

Add "demo" rake task that runs all demo:* tasks.
